requirements.txt
streamlit       ==0.82.0
Pillow          ==8.2.0
opencv-python   ==4.5.2.52
tensorflow      ==2.6.0
keras           ==2.6.0
